pyproject.toml
[tool.poetry]
name = "new-ara-api"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.7"
django = "3.1"
djangorestframework = "^3.10"
django-s3-storage = "0.13.3"
django-cors-headers = "^3.1"
django-debug-toolbar = "^2.0"
django-mysql = "^3.2"
requests = "^2.22"
beautifulsoup4 = "^4.8"
django-extensions = "^2.2"
djangorestframework-filters = "^1.0.0.dev0"
drf-yasg = "^1.17"
bleach = "^3.1"
Pillow = "^6.2"
django-environ = "^0.4.5"
gunicorn = "^19.9"
pytest = "^5.4.2"
pytest-django = "^3.9.0"
pytest-cov = "^2.8.1"
lxml = "^4.5.2"
django-crontab = "^0.7.1"
redis = "^3.5.3"
celery = {extras = ["redis"], version = "^4.4.7"}
django-celery-beat = "^2.0.0"
django-celery-results = "^1.2.1"
mysqlclient = "^2.0.1"
django-redis = "^4.12.1"
bpython = "^0.19"
django-ses = "^1.0.2"
tqdm = "^4.48.2"
cached-property = "^1.5.1"
django-cacheops = "^5.0.1"

[tool.poetry.dev-dependencies]
pylint = "^2.4"
autopep8 = "^1.4"
pylint-django = "^2.0"
django-silk = "^4.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
